Split out config::ident
