+++ src/config/mod.rs
@@ -2,6 +2,7 @@ pub mod config;
 pub mod error;
 pub mod functions;
 pub mod gobjects;
+pub mod ident;
 pub mod work_mode;
 
 pub use self::config::Config;
